fix service's nodePort change every 30s
