+++ pkg/manager/member/pd_member_manager.go
@@ -15,7 +15,6 @@ package member
 
 import (
 	"fmt"
-	"reflect"
 
 	"github.com/golang/glog"
 	"github.com/pingcap/tidb-operator/pkg/apis/pingcap.com/v1alpha1"
@@ -86,17 +85,23 @@ func (pmm *pdMemberManager) syncPDServiceForTidbCluster(tc *v1alpha1.TidbCluster
 	newSvc := pmm.getNewPDServiceForTidbCluster(tc)
 	oldSvc, err := pmm.svcLister.Services(ns).Get(controller.PDMemberName(tcName))
 	if errors.IsNotFound(err) {
+		SetServiceLastAppliedConfigAnnotation(newSvc)
 		return pmm.svcControl.CreateService(tc, newSvc)
 	}
 	if err != nil {
 		return err
 	}
 
-	if !reflect.DeepEqual(oldSvc.Spec, newSvc.Spec) {
+	equal, err := EqualService(newSvc, oldSvc)
+	if err != nil {
+		return err
+	}
+	if !equal {
 		svc := *oldSvc
 		svc.Spec = newSvc.Spec
 		// TODO add unit test
 		svc.Spec.ClusterIP = oldSvc.Spec.ClusterIP
+		SetServiceLastAppliedConfigAnnotation(&svc)
 		return pmm.svcControl.UpdateService(tc, &svc)
 	}
 
@@ -110,17 +115,23 @@ func (pmm *pdMemberManager) syncPDHeadlessServiceForTidbCluster(tc *v1alpha1.Tid
 	newSvc := pmm.getNewPDHeadlessServiceForTidbCluster(tc)
 	oldSvc, err := pmm.svcLister.Services(ns).Get(controller.PDPeerMemberName(tcName))
 	if errors.IsNotFound(err) {
+		SetServiceLastAppliedConfigAnnotation(newSvc)
 		return pmm.svcControl.CreateService(tc, newSvc)
 	}
 	if err != nil {
 		return err
 	}
 
-	if !reflect.DeepEqual(oldSvc.Spec, newSvc.Spec) {
+	equal, err := EqualService(newSvc, oldSvc)
+	if err != nil {
+		return err
+	}
+	if !equal {
 		svc := *oldSvc
 		svc.Spec = newSvc.Spec
 		// TODO add unit test
 		svc.Spec.ClusterIP = oldSvc.Spec.ClusterIP
+		SetServiceLastAppliedConfigAnnotation(newSvc)
 		return pmm.svcControl.UpdateService(tc, &svc)
 	}
 
@@ -141,7 +152,7 @@ func (pmm *pdMemberManager) syncPDStatefulSetForTidbCluster(tc *v1alpha1.TidbClu
 		return err
 	}
 	if errors.IsNotFound(err) {
-		controller.SetLastAppliedConfigAnnotation(newPDSet)
+		SetLastAppliedConfigAnnotation(newPDSet)
 		if err := pmm.setControl.CreateStatefulSet(tc, newPDSet); err != nil {
 			return err
 		}
@@ -169,7 +180,7 @@ func (pmm *pdMemberManager) syncPDStatefulSetForTidbCluster(tc *v1alpha1.TidbClu
 		}
 	}
 
-	euqal, err := controller.EqualStatefulSet(*newPDSet, *oldPDSet)
+	euqal, err := EqualStatefulSet(*newPDSet, *oldPDSet)
 	if err != nil {
 		return err
 	}
@@ -177,7 +188,7 @@ func (pmm *pdMemberManager) syncPDStatefulSetForTidbCluster(tc *v1alpha1.TidbClu
 		set := *oldPDSet
 		set.Spec.Template = newPDSet.Spec.Template
 		*set.Spec.Replicas = *newPDSet.Spec.Replicas
-		controller.SetLastAppliedConfigAnnotation(&set)
+		SetLastAppliedConfigAnnotation(&set)
 		return pmm.setControl.UpdateStatefulSet(tc, &set)
 	}
 
@@ -455,7 +466,7 @@ func (pmm *pdMemberManager) upgrade(tc *v1alpha1.TidbCluster, oldSet *apps.State
 }
 
 func (pmm *pdMemberManager) needUpgrade(tc *v1alpha1.TidbCluster, newSet *apps.StatefulSet, oldSet *apps.StatefulSet) (bool, error) {
-	same, err := controller.EqualTemplate(newSet.Spec.Template, oldSet.Spec.Template)
+	same, err := EqualTemplate(newSet.Spec.Template, oldSet.Spec.Template)
 	if err != nil {
 		return false, err
 	}
